+++ Patches/ExtendedRadioPatches.cs
@@ -3,22 +3,25 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using Colossal.IO.AssetDatabase;
-using ExtendedRadio.MonoBehaviours;
 using Game.Audio.Radio;
 using Game.SceneFlow;
 using Game.UI;
 using HarmonyLib;
 using UnityEngine;
+using BepInEx;
 using static Game.Audio.Radio.Radio;
+using System.Threading.Tasks;
+using UnityEngine.Networking;
 
 namespace ExtendedRadio.Patches
 {
 	[HarmonyPatch(typeof(GameManager), "InitializeThumbnails")]
     internal class GameManager_InitializeThumbnails
-    {
-		static readonly string IconsResourceKey = $"{MyPluginInfo.PLUGIN_NAME.ToLower()}ui";
+    {	
+		static readonly string IconsResourceKey = $"{MyPluginInfo.PLUGIN_NAME.ToLower()}";
 
 		public static readonly string COUIBaseLocation = $"coui://{IconsResourceKey}";
+
         static void Prefix(GameManager __instance)
         {		
 
@@ -51,107 +54,9 @@ namespace ExtendedRadio.Patches
 	[HarmonyPatch(typeof( Radio ), "LoadRadio")]
 	class Radio_LoadRadio {
 
-		public static GameObject gameObjectmusicLoader = new( "MusicLoader" );
-		public static MusicLoader musicLoader = gameObjectmusicLoader.AddComponent<MusicLoader>( );
-		public static string radioDirectory = Path.Combine(Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location), "CustomRadio");
-
-		public static List<string> customeRadioChannel = [];
-		public static List<string> customeNetwork = [];
-
 		static void Postfix( Radio __instance) {
 
-			Traverse radioTravers = Traverse.Create(__instance);
-
-			Dictionary<string, RadioNetwork> m_Networks = Traverse.Create(__instance).Field("m_Networks").GetValue<Dictionary<string, RadioNetwork>>();
-			Dictionary<string, RuntimeRadioChannel> m_RadioChannels = Traverse.Create(__instance).Field("m_RadioChannels").GetValue<Dictionary<string, RuntimeRadioChannel>>();
-
-			int radioNetworkIndex = m_Networks.Count()-1;
-
-			foreach(string radioNetwork in Directory.GetDirectories( radioDirectory )) {
-				if(radioNetwork != radioDirectory) {
-					if(Directory.GetFiles(radioNetwork, "*.ogg").Count() == 0) {
-
-						RadioNetwork network = new()
-						{
-							name = new DirectoryInfo(radioNetwork).Name,
-							nameId = new DirectoryInfo(radioNetwork).Name,
-							description = "A custom Network",
-							descriptionId = "A custom Network",
-							icon = File.Exists(Path.Combine(radioNetwork, "icon.svg")) ? $"{GameManager_InitializeThumbnails.COUIBaseLocation}/CustomRadio/{new DirectoryInfo(radioNetwork).Name}/icon.svg" : $"{GameManager_InitializeThumbnails.COUIBaseLocation}/resources/DefaultIcon.svg",
-							uiPriority = radioNetworkIndex++,
-							allowAds = true
-						};
-						
-						if(!m_Networks.ContainsKey(network.name)) {
-							customeNetwork.Add(network.name);
-							m_Networks.Add(network.name, network);
-						}
-						
-						foreach(string radioStation in Directory.GetDirectories( radioNetwork )) {
-							if(radioStation != radioNetwork) {
-								RadioChannel radioChannel = createRadioStation(radioStation, network.name, m_RadioChannels, radioTravers);
-								customeRadioChannel.Add(radioChannel.name);
-								m_RadioChannels.Add(radioChannel.name, radioChannel.CreateRuntime(radioStation));
-							}
-						}
-					} else {
-						RadioChannel radioChannel = createRadioStation(radioNetwork, "Public Citizen Radio", m_RadioChannels, radioTravers);
-						customeRadioChannel.Add(radioChannel.name);
-						m_RadioChannels.Add(radioChannel.name, radioChannel.CreateRuntime(radioNetwork));
-					}
-				}
-			}
-
-			radioTravers.Field("m_Networks").SetValue(m_Networks);
-			radioTravers.Field("m_RadioChannels").SetValue(m_RadioChannels);
-			radioTravers.Field("m_CachedRadioChannelDescriptors").SetValue(null);
-		}
-
-		private static RadioChannel createRadioStation( string path, string radioNetwork, Dictionary<string, RuntimeRadioChannel> m_RadioChannels, Traverse radioTravers) {
-			
-			string radioName = new DirectoryInfo(path).Name;
-			while (m_RadioChannels.ContainsKey(radioName))
-			{
-				radioName = radioName + "_" + radioTravers.Method("MakeUniqueRandomName", radioName, 4).GetValue<string>();
-			}
-			
-			AudioAsset[] audioAssets = musicLoader.LoadAllAudioClips(path, radioName , radioNetwork);
-
-            Segment segment = new()
-            {
-                type = SegmentType.Playlist,
-                clipsCap = audioAssets.Length,
-                clips = audioAssets,
-                tags = ["type:Music", "radio channel:" + radioName]
-            };
-
-            Program program = new()
-            {
-                name = "My Custom Program",
-                description = "My Custom Program",
-                icon = "coui://UIResources/Media/Radio/TheVibe.svg",
-                startTime = "00:00",
-                endTime = "00:00",
-                loopProgram = true,
-                segments = [segment]
-            };
-
-			string iconPath = $"{GameManager_InitializeThumbnails.COUIBaseLocation}/CustomRadio{(customeNetwork.Contains(radioNetwork) ? $"/{radioNetwork}" : "" )}/{radioName}/icon.svg";
-
-            RadioChannel radioChannel = new()
-            {
-                network = radioNetwork,
-                name = radioName,
-				nameId = radioName,
-                description = "A cutome Radio",
-                icon = File.Exists(Path.Combine(path, "icon.svg")) ? iconPath : $"{GameManager_InitializeThumbnails.COUIBaseLocation}/resources/DefaultIcon.svg", //"Media/Radio/Stations/TheVibe.svg";
-                uiPriority = 1,
-                programs = [program]
-            };
-
-			musicLoader.AddToDataBase(radioChannel);
-
-            return radioChannel;
+			ExtendedRadio.OnLoadRadio(__instance);
 
 		}
 	}
@@ -161,9 +66,9 @@ namespace ExtendedRadio.Patches
 	{
 		static bool Prefix( Radio __instance, RuntimeSegment segment)
 		{		
-			if(Radio_LoadRadio.customeRadioChannel.Contains(__instance.currentChannel.name)) {
+			if(ExtendedRadio.customeRadioChannelsName.Contains(__instance.currentChannel.name)) {
 
-				IEnumerable<AudioAsset> assets = Radio_LoadRadio.musicLoader.GetAudiAssets(__instance, segment.type);
+				IEnumerable<AudioAsset> assets = ExtendedRadio.GetAudiAssetsFromAudioDataBase(__instance, segment.type);
 				List<AudioAsset> list = [.. assets];
 				System.Random rnd = new();
 				List<int> list2 = (from x in Enumerable.Range(0, list.Count)
@@ -180,6 +85,40 @@ namespace ExtendedRadio.Patches
 				return false;
 			}
 			return true;
+		}
+	}
+
+	[HarmonyPatch(typeof(AudioAsset), "LoadAsync")]
+	internal class AudioAssetLoadAsyncPatch
+	{
+		static bool Prefix(AudioAsset __instance, ref Task<AudioClip> __result)
+		{	
+			if(!ExtendedRadio.customeRadioChannelsName.Contains(__instance.GetMetaTag(AudioAsset.Metatag.RadioChannel))) return true;
+			
+			__result = LoadAudioFile(__instance);
+			return false;
+		}
+
+		private static async Task<AudioClip> LoadAudioFile(AudioAsset audioAsset)
+		{
+			Traverse audioAssetTravers = Traverse.Create(audioAsset);
+
+			if(audioAssetTravers.Field("m_Instance").GetValue() == null)
+			{
+				string sPath = ExtendedRadio.GetClipPathFromAudiAsset(audioAsset);
+				using UnityWebRequest www = UnityWebRequestMultimedia.GetAudioClip("file://" + sPath, AudioType.OGGVORBIS);
+				((DownloadHandlerAudioClip) www.downloadHandler).streamAudio = true;
+				await www.SendWebRequest();
+				AudioClip clip = DownloadHandlerAudioClip.GetContent(www);
+				www.Dispose();
+
+				clip.name = sPath;
+				clip.hideFlags = HideFlags.DontSave;
+
+				audioAssetTravers.Field("m_Instance").SetValue(clip);
+			}
+
+			return (AudioClip) audioAssetTravers.Field("m_Instance").GetValue();
 		}
 	}
 }